fix: names and create a interface for service layer
